+++ lib/legion/extensions/mesh/transport/messages/preference_query.rb
@@ -1,0 +1,34 @@
+# frozen_string_literal: true
+
+module Legion
+  module Extensions
+    module Mesh
+      module Transport
+        module Messages
+          class PreferenceQuery < Legion::Transport::Message
+            def exchange
+              Legion::Transport::Exchanges::Agent
+            end
+
+            def routing_key
+              "agent.#{@options[:target_agent_id]}"
+            end
+
+            def message
+              {
+                type:                'preference_query',
+                requesting_agent_id: @options[:requesting_agent_id],
+                domains:             @options[:domains] || [],
+                requested_at:        Time.now.to_s
+              }
+            end
+
+            def type
+              'preference_query'
+            end
+          end
+        end
+      end
+    end
+  end
+end